Split timeline ops off blobdb
