+++ rockworkd/libpebble/blobdb.h
@@ -8,7 +8,6 @@
 #include "appmetadata.h"
 
 #include <QObject>
-#include <QDateTime>
 
 class BlobDB : public QObject
 {
@@ -37,13 +36,6 @@ public:
 
     explicit BlobDB(Pebble *pebble, WatchConnection *connection);
 
-    void insertNotification(const Notification &notification);
-    void insertTimelinePin(const QUuid &uuid, TimelineItem::Layout layout, bool isAllDay, const QDateTime &startTime, const QDateTime &endTime, const QString &title, const QString &description, const QMap<QString, QString> fields, bool recurring);
-    void removeTimelinePin(const QUuid &uuid);
-    void insertReminder(const QUuid &uuid, const QUuid &parentId, const QString &title, const QString &subtitle, const QString &body, const QDateTime &remindTime);
-    void clearTimeline();
-    void syncCalendar(const QList<CalendarEvent> &events);
-
     void clearApps();
     void insertAppMetaData(const AppInfo &info, const bool force=false);
     void removeApp(const AppInfo &info);
@@ -57,14 +49,9 @@ public:
 
 private slots:
     void blobCommandReply(const QByteArray &data);
-    void actionInvoked(const QByteArray &data);
-    void sendActionReply();
     void sendNext();
 
 signals:
-    void muteSource(const QString &sourceId);
-    void removeNotification(const QUuid &uuid);
-    void actionTriggered(const QUuid &uuid, const QString &actToken);
     void appInserted(const QUuid &uuid);
 
 private:
@@ -88,11 +75,6 @@ private:
 
     Pebble *m_pebble;
     WatchConnection *m_connection;
-
-    QHash<QUuid, Notification> m_notificationSources;
-
-    QList<CalendarEvent> m_calendarEntries;
-    CalendarEvent findCalendarEvent(const QString &id);
 
     HealthParams m_healthParams;
